+++ src/test/python_tests/lsp_test_client/utils.py
@@ -9,6 +9,11 @@ import os
 import pathlib
 import platform
 import random
+from typing import Any, List
+
+
+import lsprotocol.converters as cv
+import lsprotocol.types as lsp
 
 from .constants import PROJECT_ROOT
 
@@ -66,3 +71,24 @@ def get_initialization_options():
     setting["cwd"] = str(PROJECT_ROOT)
 
     return {"settings": [setting], "globalSettings": setting}
+
+
+def apply_text_edits(text: str, text_edits: List[lsp.TextEdit]) -> str:
+    if not text_edits:
+        return text
+
+    offsets = [0]
+    for line in text.splitlines(keepends=True):
+        offsets.append(offsets[-1] + len(line))
+
+    for text_edit in reversed(text_edits):
+        start_offset = offsets[text_edit.range.start.line] + text_edit.range.start.character
+        end_offset = offsets[text_edit.range.end.line] + text_edit.range.end.character
+        text = text[:start_offset] + text_edit.new_text + text[end_offset:]
+    return text
+
+
+def destructure_text_edits(text_edits: List[Any]) -> List[lsp.TextEdit]:
+    """Converts text edits from the language server to the format used by the test client."""
+    converter = cv.get_converter()
+    return [converter.structure(text_edit, lsp.TextEdit) for text_edit in text_edits]
